Update README.md
The ISS Tracker and Notifier is a Python script designed to keep you informed about the whereabouts of the International Space Station (ISS) in real-time. It uses the Open Notify API to track the ISS's current latitude and longitude, allowing you to know when it's passing near your location. Moreover, the script checks if it's nighttime at your location using the Sunrise-Sunset API, ensuring that you receive email notifications only when the ISS is overhead during nighttime hours. Stay connected to the marvels of space exploration with this simple yet powerful tracking tool.